Fix blurry avatars by fetching at bigger size
Instead of fetching the images at 26x26, it's fetching at 88x88. (the
image in a notification seems to be 44x44, so 88x88 should cover the
case of Retina displays)

That would mean that the cache should end up with more stuff in it since
it'd get the avatars from the notifications at 88x88 and the avatar for
the dropdown menu at 26x26... Not sure of the best way to do it though.

Hope that helps.